+++ parser/src/tests/blocks/mod.rs
@@ -1,6 +1,6 @@
 mod block;
 mod compound_delimited;
-mod r#macro;
+mod media;
 mod raw_delimited;
 mod section;
 mod simple;
