+++ parser/src/tests/fixtures/blocks/mod.rs
@@ -4,8 +4,8 @@ pub(crate) use block::TBlock;
 mod compound_delimited;
 pub(crate) use compound_delimited::TCompoundDelimitedBlock;
 
-mod r#macro;
-pub(crate) use r#macro::TMediaBlock;
+mod media;
+pub(crate) use media::TMediaBlock;
 
 mod raw_delimited;
 pub(crate) use raw_delimited::TRawDelimitedBlock;
